# Update "show campaign" interface to use GraphQL.