--- a/resources/assets/app.js
+++ b/resources/assets/app.js
@@ -8,7 +8,6 @@ import mountContainer from './utilities/MountContainer';
 
 import Signup from './components/Signup';
 import Application from './components/Application';
-import CampaignSingle from './components/CampaignSingle';
 import CampaignIdSingle from './components/CampaignIdSingle';
 
 // Display environment badge on local, dev, or QA:
@@ -23,7 +22,6 @@ ready(() => {
   // For "legacy" pages that render using our custom helpers:
   mountContainer({
     CampaignIdSingle,
-    CampaignSingle,
     Signup,
   });
 });