--- a/resources/assets/components/ReviewablePost.js
+++ b/resources/assets/components/ReviewablePost.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import { map } from 'lodash';
 import gql from 'graphql-tag';
 import usePortal from 'react-useportal';
 import { Link } from 'react-router-dom';
@@ -6,11 +7,12 @@ import { parse, format } from 'date-fns';
 
 import Quantity from './Quantity';
 import PostTile from './PostTile';
+import { TAGS } from '../helpers';
 import TextBlock from './TextBlock';
 import Modal from './utilities/Modal';
 import PostCard from './utilities/PostCard';
 import DeleteButton from './utilities/DeleteButton';
-import { AllTagButtons, TagButtonFragment } from './utilities/TagButton';
+import TagButton, { TagButtonFragment } from './utilities/TagButton';
 import QuantityForm, { QuantityFormFragment } from './utilities/QuantityForm';
 import MetaInformation from './MetaInformation';
 import {
@@ -163,7 +165,9 @@ const ReviewablePost = ({ post }) => {
                 (see definitions)
               </a>
             </h4>
-            <AllTagButtons post={post} />
+            {map(TAGS, (label, id) => (
+              <TagButton key={id} tag={id} label={label} post={post} />
+            ))}
           </div>
         ) : null}
         <div className="container__row">