--- a/resources/assets/components/utilities/TagButton.js
+++ b/resources/assets/components/utilities/TagButton.js
@@ -21,7 +21,7 @@ const TAG_POST_MUTATION = gql`
   }
 `;
 
-export const TagButton = ({ post, tag, label }) => {
+const TagButton = ({ post, tag, label }) => {
   const hasTag = post.tags.includes(tag);
 
   const [tagPost] = useMutation(TAG_POST_MUTATION, {
@@ -50,26 +50,5 @@ export const TagButton = ({ post, tag, label }) => {
     </button>
   );
 };
-
-export const AllTagButtons = ({ post }) => (
-  <>
-    <TagButton post={post} tag="good-submission" label="Good Submission" />
-    <TagButton post={post} tag="good-quote" label="Good Quote" />
-    <TagButton post={post} tag="good-for-brand" label="Good For Brand" />
-    <TagButton post={post} tag="good-for-sponsor" label="Good For Sponsor" />
-    <TagButton post={post} tag="group-photo" label="Group Photo" />
-    <TagButton post={post} tag="hide-in-gallery" label="Hide In Gallery 👻" />
-    <TagButton post={post} tag="irrelevant" label="Irrelevant" />
-    <TagButton post={post} tag="inappropriate" label="Inappropriate" />
-    <TagButton
-      post={post}
-      tag="unrealistic-quantity"
-      label="Unrealistic Quantity"
-    />
-    <TagButton post={post} tag="test" label="Test" />
-    <TagButton post={post} tag="incomplete-action" label="Incomplete Action" />
-    <TagButton post={post} tag="bulk" label="Bulk" />
-  </>
-);
 
 export default TagButton;