… resources/assets/components/CampaignSingle/index.js