@DFurnes 2019-11-12
Something that doesn't show up in the screenshot or code diff but seems worth highlighting... building these interfaces via GraphQL is _screamingly_ faster than querying them from our PHP backend (mainly because we can efficiently batch up includes):

| Page            | Before | After | Comparison  |
|-----------------|--------|-------|-------------|
| `7950/pending`  | 4.22s  | 0.95s | 4.5x faster |
| `1144/accepted` | 3.79s  | 1.28s | 3.0x faster |
| `9013/pending`  | 2.71s  | 0.87s | 3.1x faster |
| `8026/accepted` | 2.53s  | 0.51s | 4.9x faster |
| `7/pending`     | 5.27s  | 0.74s | 7.1x faster |
| `7/accepted`    | 9.22s  | 1.23s | 7.5x faster |